# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Documentation for API e-mosque endpoints",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/kegiatan": {
            "post": {
                "summary": "Create a new kegiatan",
                "description": "Create a new kegiatan with the provided details.",
                "operationId": "createKegiatan",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_kegiatan",
                                    "nama_pengaju",
                                    "deskripsi",
                                    "tgl_kegiatan",
                                    "id_pj"
                                ],
                                "properties": {
                                    "id_kegiatan": {
                                        "description": "ID of the kegiatan",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "nama_pengaju": {
                                        "description": "Name of the pengaju",
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "deskripsi": {
                                        "description": "Description of the kegiatan",
                                        "type": "string",
                                        "example": "Deskripsi kegiatan"
                                    },
                                    "tgl_kegiatan": {
                                        "description": "Date of the kegiatan",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2022-05-25"
                                    },
                                    "id_pj": {
                                        "description": "ID of the PJ",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful creation of kegiatan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Status of the response",
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "description": "Message indicating the status of the kegiatan creation",
                                            "type": "string",
                                            "example": "berhasil menambah data kegiatan"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Status of the response",
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "description": "Message indicating the status of the kegiatan creation failure",
                                            "type": "string",
                                            "example": "gagal menambah data kegiatan"
                                        },
                                        "data": {
                                            "description": "Additional error data",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/kegiatan/{id_kegiatan}": {
            "put": {
                "summary": "Update kegiatan",
                "description": "Update kegiatan data by ID.",
                "operationId": "updateKegiatan",
                "parameters": [
                    {
                        "name": "id_kegiatan",
                        "in": "path",
                        "description": "ID of the kegiatan to be updated",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nama_pengaju",
                                    "deskripsi",
                                    "tgl_kegiatan",
                                    "id_pj"
                                ],
                                "properties": {
                                    "nama_pengaju": {
                                        "description": "Name of the pengaju",
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "deskripsi": {
                                        "description": "Description of the kegiatan",
                                        "type": "string",
                                        "example": "Deskripsi kegiatan"
                                    },
                                    "tgl_kegiatan": {
                                        "description": "Date of the kegiatan",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2022-05-25"
                                    },
                                    "id_pj": {
                                        "description": "ID of the PJ",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful update of kegiatan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Status of the response",
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "description": "Message indicating the status of the kegiatan update",
                                            "type": "string",
                                            "example": "berhasil memperbarui data kegiatan"
                                        },
                                        "data": {
                                            "description": "Updated kegiatan data",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Kegiatan data not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Status of the response",
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "description": "Message indicating the status of the kegiatan update failure",
                                            "type": "string",
                                            "example": "id kegiatan tidak ditemukan"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "summary": "Delete kegiatan",
                "description": "Delete kegiatan data by ID.",
                "operationId": "deleteKegiatan",
                "parameters": [
                    {
                        "name": "id_kegiatan",
                        "in": "path",
                        "description": "ID of the kegiatan to be deleted",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful deletion of kegiatan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Status of the response",
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "description": "Message indicating the status of the kegiatan deletion",
                                            "type": "string",
                                            "example": "kegiatan berhasil dihapus"
                                        },
                                        "data": {
                                            "description": "No data returned",
                                            "type": "null"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Kegiatan data not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Status of the response",
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "description": "Message indicating the status of the kegiatan deletion failure",
                                            "type": "string",
                                            "example": "Data kegiatan tidak ditemukan"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}